+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Include location-count on libraries page. Refs UIORG-66.
 * Refresh lookup tables on mount. Refs UIORG-69.
 * Refactoring away structures. Refs STCOM-277.
+* Refresh location counts on mount. Refs UIORG-60, UIORG-63, UIORG-66.
 
 ## [2.2.0](https://github.com/folio-org/ui-organization/tree/v2.2.0) (2017-09-01)
 [Full Changelog](https://github.com/folio-org/ui-organization/compare/v2.1.0...v2.2.0)